requirements.txt
plotly==5.1.0
streamlit==0.86.0
numpy==1.19.5
tensorflow-cpu==2.5.0
boto3==1.18.16
psycopg2==2.9.1
Pillow==8.3.1
python-dotenv==0.19.0
scikit_learn==0.24.2
